Add ObjectDataAPI and rename package for FunctionAndMethodAPI
