+++ src/main/java/io/github/ngspace/hudder/compilers/abstractions/AScriptingLanguageCompiler.java
@@ -9,10 +9,10 @@ import io.github.ngspace.hudder.Hudder;
 import io.github.ngspace.hudder.compilers.utils.CompileException;
 import io.github.ngspace.hudder.compilers.utils.HudInformation;
 import io.github.ngspace.hudder.compilers.utils.IScriptingLanguageEngine;
-import io.github.ngspace.hudder.compilers.utils.unifiedcomp.ArrayElementManager;
+import io.github.ngspace.hudder.compilers.utils.functionandmethodapi.ArrayElementManager;
 import io.github.ngspace.hudder.data_management.BooleanData;
 import io.github.ngspace.hudder.data_management.NumberData;
-import io.github.ngspace.hudder.data_management.ObjectData;
+import io.github.ngspace.hudder.data_management.ObjectDataAPI;
 import io.github.ngspace.hudder.data_management.StringData;
 import io.github.ngspace.hudder.main.HudCompilationManager;
 import io.github.ngspace.hudder.main.config.HudderConfig;
@@ -94,7 +94,7 @@ public abstract class AScriptingLanguageCompiler extends AVarTextCompiler {
 		if ( obj!=null) return obj;
 		if ((obj=StringData.getString (key))!=null) return obj;
 		if ((obj=BooleanData.getBoolean(key))!=null) return obj;
-		if ((obj=ObjectData.getObject(key))!=null) return obj;
+		if ((obj=ObjectDataAPI.getObject(key))!=null) return obj;
 		if ((obj=get(key))!=null) return obj;
 		if ((obj=Hudder.config.globalVariables.get(key))!=null) return obj;
 		return null;
