+++ src/main/java/io/github/ngspace/hudder/compilers/abstractions/AV2Compiler.java
@@ -6,11 +6,11 @@ import java.util.Map;
 import io.github.ngspace.hudder.compilers.utils.CompileException;
 import io.github.ngspace.hudder.compilers.utils.CompileState;
 import io.github.ngspace.hudder.compilers.utils.HudInformation;
-import io.github.ngspace.hudder.compilers.utils.unifiedcomp.FunctionAndMethodAPI;
-import io.github.ngspace.hudder.compilers.utils.unifiedcomp.FunctionAndMethodAPI.BindableConsumer;
-import io.github.ngspace.hudder.compilers.utils.unifiedcomp.FunctionAndMethodAPI.BindableFunction;
-import io.github.ngspace.hudder.compilers.utils.unifiedcomp.FunctionAndMethodAPI.ConsumerBinder;
-import io.github.ngspace.hudder.compilers.utils.unifiedcomp.FunctionAndMethodAPI.FunctionBinder;
+import io.github.ngspace.hudder.compilers.utils.functionandmethodapi.FunctionAndMethodAPI;
+import io.github.ngspace.hudder.compilers.utils.functionandmethodapi.FunctionAndMethodAPI.BindableConsumer;
+import io.github.ngspace.hudder.compilers.utils.functionandmethodapi.FunctionAndMethodAPI.BindableFunction;
+import io.github.ngspace.hudder.compilers.utils.functionandmethodapi.FunctionAndMethodAPI.ConsumerBinder;
+import io.github.ngspace.hudder.compilers.utils.functionandmethodapi.FunctionAndMethodAPI.FunctionBinder;
 import io.github.ngspace.hudder.main.HudCompilationManager;
 import io.github.ngspace.hudder.main.config.HudderConfig;
 import io.github.ngspace.hudder.v2runtime.V2Runtime;
