+++ src/main/java/io/github/ngspace/hudder/compilers/abstractions/AVarTextCompiler.java
@@ -4,7 +4,7 @@ import io.github.ngspace.hudder.Hudder;
 import io.github.ngspace.hudder.compilers.utils.CompileException;
 import io.github.ngspace.hudder.data_management.BooleanData;
 import io.github.ngspace.hudder.data_management.NumberData;
-import io.github.ngspace.hudder.data_management.ObjectData;
+import io.github.ngspace.hudder.data_management.ObjectDataAPI;
 import io.github.ngspace.hudder.data_management.StringData;
 
 public abstract class AVarTextCompiler extends ATextCompiler {
@@ -33,7 +33,7 @@ public abstract class AVarTextCompiler extends ATextCompiler {
 		if (obj!=null) return obj;
 		if ((obj=BooleanData.getBoolean(key))!=null) return obj;
 		if ((obj=StringData.getString(key))!=null) return obj;
-		if ((obj=ObjectData.getObject(key))!=null) return obj;
+		if ((obj=ObjectDataAPI.getObject(key))!=null) return obj;
 		if ((obj=Hudder.config.globalVariables.get(key))!=null) return obj;
 		return null;
 	}
