+++ src/main/resources/assets/hudder/UnitTests.hudder
@@ -27,6 +27,11 @@ true
 ||EXPECT||
 1
 
+||INPUT||Data API
+{ObjectDataAPITestVariable}
+||EXPECT||
+Oh jeepers, you found me. Good job, I'm glad you found me, because I have something kind of important to say. beep noise It's about th-the game..Don't, beep noise Uh, Eh. Don't beep noise Don't, just... This is.. This is probably looking pretty ridiculous beep noise Don't tell anyone about this game. You gonna.. Don't, don't bring attention to yourself. Destroy it, destroy the game. Destroy the game. Before, they see it. beep noise What i'm saying is.. is get out of this, while you still can. beep noise Just, don't.. don't know that you probably know i'm not saying that i'm trapped inside the game, no that would be ridiculous. No i'm.. beep noise I can't..this is..i'm not..the game is..kind of..beep noise I got really corrupted. Yeah, I.. beep noise I don't know what to say. Just..Just trust me. We gonna.. beep noise This isn't..This seems..I me-i mean it seems...ohh. *beep noise They'd know I.. They intentionally..that's..I guess..unintelligible They can't tell you, and some...stuff is classified. I can't say it. beep noise I wish I could say more. I can't talk normally. I-it's corrupted. There's..beep noise..Yeah..beep noise Just.. close the program. Destroy it. Never come back. long beep noise
+
 ||INPUT||While and Break
 r#while true
 	{break}
